+++ packages/benchmark.js-plugin/package.json
@@ -33,14 +33,12 @@
     "@babel/preset-env": "^7.22.5",
     "@types/benchmark": "^2.1.2",
     "@types/lodash": "^4.14.195",
-    "@types/stack-trace": "^0.0.30",
     "benchmark": "^2.1.4",
     "jest-mock-extended": "^3.0.4"
   },
   "dependencies": {
     "@codspeed/core": "workspace:^5.4.0",
-    "lodash": "^4.17.10",
-    "stack-trace": "1.0.0-pre2"
+    "lodash": "^4.17.10"
   },
   "peerDependencies": {
     "benchmark": "^2.1.0"
