+++ packages/core/package.json
@@ -33,6 +33,7 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@types/find-up": "^4.0.0",
+    "@types/stack-trace": "^0.0.30",
     "node-addon-api": "^5.1.0",
     "node-gyp": "^12.2.0",
     "openapi-typescript-codegen": "^0.23.0",
@@ -42,6 +43,7 @@
     "axios": "^1.4.0",
     "find-up": "^6.3.0",
     "form-data": "^4.0.4",
-    "node-gyp-build": "^4.6.0"
+    "node-gyp-build": "^4.6.0",
+    "stack-trace": "1.0.0-pre2"
   }
 }
